+++ src/heart/renderers/vibe/__init__.py
@@ -1,0 +1,2 @@
+from heart.renderers.vibe.renderer import VibeScene as VibeScene
+from heart.renderers.vibe.state import VibeState as VibeState
